:recycle: Refactored icon typing across custom icons
