+++ src/components/Icons/Fallback.tsx
@@ -1,8 +1,8 @@
 import { FC } from 'react';
 
-import { AppIconProps } from 'src/types/AppIcon';
+import { SizeIconProps } from 'src/types/Icon';
 
-const Fallback: FC<AppIconProps> = ({ size }) => (
+const Fallback: FC<SizeIconProps> = ({ size }) => (
   <svg
     width={size ? size : 48}
     height={size ? size : 48}
