+++ src/components/Icons/FeedBackIcon/PositiveFilled.tsx
@@ -1,8 +1,8 @@
 import { FC } from 'react';
 
-import { AppIconProps } from 'src/types/AppIcon';
+import { SizeIconProps } from 'src/types/Icon';
 
-const PositiveFilled: FC<AppIconProps> = ({ size }) => (
+const PositiveFilled: FC<SizeIconProps> = ({ size }) => (
   <svg
     width={size ? size : 48}
     height={size ? size : 48}
