+++ src/components/Icons/FeedBackIcon/PositiveOutlined.tsx
@@ -1,8 +1,8 @@
 import { FC } from 'react';
 
-import { AppIconProps } from 'src/types/AppIcon';
+import { SizeIconProps } from 'src/types/Icon';
 
-const PositiveOutlined: FC<AppIconProps> = ({ size }) => (
+const PositiveOutlined: FC<SizeIconProps> = ({ size }) => (
   <svg
     width={size ? size : 48}
     height={size ? size : 48}
