+++ src/types/index.ts
@@ -1,4 +1,4 @@
-export * from './AppIcon';
+export * from 'src/types/Icon';
 export * from './Environment';
 export * from './Errors';
 export * from './Field';
